moved getTiles and filtering to utils
