+++ modules/lib/index.js
@@ -1,3 +1,2 @@
 export { d3combobox } from './d3.combobox';
-export { d3geoTile } from './d3.geo.tile';
 export { d3keybinding } from './d3.keybinding';
