+++ modules/services/streetside.js
@@ -17,7 +17,6 @@ import {
 import rbush from 'rbush';
 import { t } from '../util/locale';
 import { jsonpRequest } from '../util/jsonp_request';
-import { d3geoTile as d3_geoTile } from '../lib/d3.geo.tile';
 
 import {
     geoExtent,
@@ -29,9 +28,11 @@ import {
 } from '../geo';
 
 import { utilDetect } from '../util/detect';
-import { utilQsString, utilRebind } from '../util';
+import { utilQsString, utilRebind, utilTile } from '../util';
 
 import Q from 'q';
+
+var geoTile = utilTile();
 
 var bubbleApi = 'https://dev.virtualearth.net/mapcontrol/HumanScaleServices/GetBubbles.ashx?';
 var streetsideImagesApi = 'https://t.ssl.ak.tiles.virtualearth.net/tiles/';
@@ -85,46 +86,6 @@ function localeTimestamp(s) {
     return d.toLocaleString(detected.locale, options);
 }
 
-/**
- * getTiles() returns array of d3 geo tiles.
- * Using d3.geo.tiles.js from lib, gets tile extents for each grid tile in a grid created from
- * an area around (and including) the current map view extents.
- */
-function getTiles(projection, margin) {
-    // s is the current map scale
-    // z is the 'Level of Detail', or zoom-level, where Level 1 is far from the earth, and Level 23 is close to the ground.
-    // ts ('tile size') here is the formula for determining the width/height of the map in pixels, but with a modification.
-    // See 'Ground Resolution and Map Scale': //https://msdn.microsoft.com/en-us/library/bb259689.aspx.
-    // As used here, by subtracting constant 'tileZoom' from z (the level), you end up with a much smaller value for the tile size (in pixels).
-    var s = projection.scale() * 2 * Math.PI;
-    var z = Math.max(Math.log(s) / Math.log(2) - 8, 0);
-    var ts = 256 * Math.pow(2, z - tileZoom);
-    var origin = [
-        s / 2 - projection.translate()[0],
-        s / 2 - projection.translate()[1]
-    ];
-
-    var tiler = d3_geoTile()
-        .scaleExtent([tileZoom, tileZoom])
-        .scale(s)
-        .size(projection.clipExtent()[1])
-        .translate(projection.translate())
-        .margin(margin || 0);   // request nearby tiles so we can connect sequences.
-
-    return tiler()
-        .map(function(tile) {
-            var x = tile[0] * ts - origin[0];
-            var y = tile[1] * ts - origin[1];
-            return {
-                id: tile.toString(),
-                xyz: tile,
-                extent: geoExtent(
-                    projection.invert([x, y + ts]),
-                    projection.invert([x + ts, y])
-                )
-            };
-        });
-}
 
 /**
  * loadTiles() wraps the process of generating tiles and then fetching image points for each tile.
@@ -133,10 +94,7 @@ function loadTiles(which, url, projection, margin) {
     var s = projection.scale() * 2 * Math.PI;
     var currZoom = Math.floor(Math.max(Math.log(s) / Math.log(2) - 8, 0));
 
-    // breakup the map view into tiles
-    var tiles = getTiles(projection, margin).filter(function (t) {
-        return !nearNullIsland(t.xyz[0], t.xyz[1], t.xyz[2]);
-    });
+    var tiles = geoTile.filterNullIsland(geoTile.getTiles(tileZoom, projection, margin));
 
     tiles.forEach(function (tile) {
        loadNextTilePage(which, currZoom, url, tile);
