+++ package.json
@@ -66,7 +66,7 @@
     "json-stringify-pretty-compact": "^1.1.0",
     "jsonschema": "^1.1.0",
     "mapillary-js": "2.12.1",
-    "mapillary_sprite_source": "^1.4.0",
+    "mapillary_sprite_source": "^1.5.0",
     "minimist": "^1.2.0",
     "mocha": "^5.0.0",
     "mocha-phantomjs-core": "^2.1.0",
